RunFitManager -> FitWorkerLauncher (files renaming)
