+++ GUI/coregui/Views/FitWidgets/FitWorkerLauncher.h
